+++ src/features/identityCheck/pages/phoneValidation/errors/tests/PhoneValidationTooManyAttempts.web.test.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import waitForExpect from 'wait-for-expect'
 
 import { navigate } from '__mocks__/@react-navigation/native'
-import { PhoneValidationTooManyAttempts } from 'features/auth/signup/PhoneValidation/PhoneValidationTooManyAttempts'
 import { contactSupport } from 'features/auth/support.services'
+import { PhoneValidationTooManyAttempts } from 'features/identityCheck/pages/phoneValidation/errors/PhoneValidationTooManyAttempts'
 import * as NavigationHelpers from 'features/navigation/helpers/openUrl'
 import { homeNavConfig } from 'features/navigation/TabBar/helpers'
 import { fireEvent, render } from 'tests/utils/web'
